:truck: Moved types, annotated types and enums to types
