+++ data_safe_haven/types/annotated_types.py
@@ -19,7 +19,7 @@ AzureShortName = Annotated[str, Field(min_length=1, max_length=24)]
 AzureLongName = Annotated[str, Field(min_length=1, max_length=64)]
 AzureLocation = Annotated[str, AfterValidator(validate_azure_location)]
 AzureVmSku = Annotated[str, AfterValidator(validate_azure_vm_sku)]
-EmailAdress = Annotated[str, AfterValidator(validate_email_address)]
+EmailAddress = Annotated[str, AfterValidator(validate_email_address)]
 Fqdn = Annotated[str, AfterValidator(validate_fqdn)]
 Guid = Annotated[str, AfterValidator(validate_aad_guid)]
 IpAddress = Annotated[str, AfterValidator(validate_ip_address)]
