+++ data_safe_haven/utility/__init__.py
@@ -1,21 +1,12 @@
-from .annotated_types import AzureLocation, AzureLongName, Guid
 from .directories import config_dir
-from .enums import DatabaseSystem, SoftwarePackageCategory
 from .file_reader import FileReader
 from .logger import LoggingSingleton, NonLoggingSingleton
 from .singleton import Singleton
-from .types import PathType
 
 __all__ = [
-    "AzureLocation",
-    "AzureLongName",
     "config_dir",
-    "DatabaseSystem",
     "FileReader",
-    "Guid",
     "LoggingSingleton",
     "NonLoggingSingleton",
-    "PathType",
     "Singleton",
-    "SoftwarePackageCategory",
 ]
